+++ src/main/java/pl/asie/lib/chat/ChatHandler.java
@@ -20,10 +20,10 @@ import java.util.HashMap;
 import java.util.List;
 
 public class ChatHandler {
-	private HashMap<String, String> actions = new HashMap<String, String>();
-	public boolean enableChatFeatures, enableShout, enableGreentext, enableColor;
-	public int CHAT_RADIUS, nickLevel;
-	public String colorAction, messageFormat, shoutPrefix;
+	private final HashMap<String, String> actions = new HashMap<String, String>();
+	public final boolean enableChatFeatures, enableShout, enableGreentext, enableColor;
+	public final int CHAT_RADIUS, nickLevel, realnameLevel;
+	public final String colorAction, messageFormat, shoutPrefix;
 
 	public ChatHandler(Configuration config) {
 		CHAT_RADIUS = config.get("chat", "chatRadius", 0).getInt();
@@ -33,6 +33,7 @@ public class ChatHandler {
 		enableColor = config.get("base", "enableColor", true).getBoolean(true);
 		enableGreentext = config.get("chat", "enableGreentext", false, ">implying anyone will ever turn this on").getBoolean(false);
 		nickLevel = config.get("chat", "nicknamesForEveryone", true, "Disable to make changing the own nickname require Op rights on a server").getBoolean(true) ? 0 : 2;
+		realnameLevel =config.get("chat", "realnamesForEveryone", true, "Disable to make looking up the real name of others require Op rights on a server").getBoolean(true) ? 0 : 2;
 		colorAction = config.get("chat", "colorMe", "5").getString();
 		messageFormat = config.get("chat", "formatMessage", "<%u> %m", "%u - username; %m - message; %w - dimension; %H - hours; %M - minutes; %S - seconds").getString();
 	}