+++ src/main/java/pl/asie/lib/chat/NicknameRepository.java
@@ -13,6 +13,7 @@ import net.minecraftforge.common.DimensionManager;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import pl.asie.lib.api.chat.INicknameHandler;
 import pl.asie.lib.api.chat.INicknameRepository;
+import pl.asie.lib.util.ChatUtils;
 import pl.asie.lib.util.PlayerUtils;
 
 import java.io.File;
@@ -62,8 +63,15 @@ public class NicknameRepository implements INicknameRepository {
 	}
 	
 	public String getUsername(String nickname) {
-		BiMap<String, String> usernames = nicknames.inverse();
-		return usernames.containsKey(nickname) ? usernames.get(nickname) : null;
+		//BiMap<String, String> usernames = nicknames.inverse();
+		for(String username : nicknames.keySet()) {
+			String nick = nicknames.get(username);
+			if(nick != null && ChatUtils.stripColors(nick).equals(nickname)) {
+				return username;
+			}
+		}
+		//return usernames.containsKey(nickname) ? usernames.get(nickname) : null;
+		return null;
 	}
 	
 	public String getNickname(String username) {
@@ -72,7 +80,7 @@ public class NicknameRepository implements INicknameRepository {
 	
 	public String getRawNickname(String username) {
 		String nn = getNickname(username);
-		return nn.replaceAll("&.", "");
+		return ChatUtils.stripColors(nn);
 	}
 
 	public boolean hasNickname(String username) {